Add grid layout to entire page and nav sidebar
Rearranged the nav and header divs so that the grid layout would work
better (probably didn't need to). Using a pretty simple 2-column grid
for the main body, with the left-hand nav bar spanning the entire page.
The nav bar's internal grid is also simple - 2 columns, 12 rows, with 2
empty sets of divs to create the spacing. I may take those out and just
make two of the rows taller than the others. We'll see.