Apply grid to header area
This is getting a little trickier. I've set up grids inside grids. The
entire header area is a grid container. Then, I've set the search area
as it's own grid with four columns (I have the magnifying glass in the
same div as the search input as its label). Then, what I'm calling the
"controls" header is its own grid with just 2 columns. The left side of
that header section, the avatar, greeting, and name (which I'm calling)
"greeting" is another grid, 2x2, with the avatar spanning both rows.
The buttons on the right are another grid (1 row, 3 cols).

I think I can actually simplify this by putting all of that section in
a 2x2 grid, and then using 1 nested grid in each cell (other than the
search).