Style the header and subsections
I'm not overly happy with the way I styled the "user info" (or
"profile") section of the header. I ended up using static sizes, which
is fine for the notification icon and the avatar, but if the name
changes, that cell should grow or shrink. Or maybe not shrink, but it
feels like it should allow for a longer name. I should probably try
min or minmax there.