Tweak layout of Header
After a night to think on the spacing, I went back and looked at my
options, and stumbled on "min-content" and "max-content". I left the
column sizes static for the bell icon and the small avatar, but set the
name column to "max-content", allowing it to expand as the name length
increases (the one issue is that if the name gets too long, that nested
grid could infringe on the search bar, so I should look into a max or
maybe a minmax setting). I then set the button grids to "min-content",
as they don't tracks any wider than they are (I set the widths, so I
could have used that number, but this allows some future adjustments to
the buttons without having to also change the grid).