Set some styles for cards
While setting the styles for the cards, I realized that I may need to
redsign the main body grid to make it easier to actuall lay out. At the
moment, the Projects head is in the cards grid, which makes it harder
to lay out the cards, as I have less control over the track sizes.