Set styles for cards
Cards should be nicely responsive, now, as I am getting a handle on the
auto-fill property for grid. They should shrink down to 30ch wide and
then wrap and expand back to 1fr. I set something similar for the
height, so the content shouldn't overflow the bottom.

I also added some commented out code that will be needed if the page is
to be full viewport height. Some things need to be tweaked for that,
but I don't know if it's needed, so the lines are commented out for now.