Set styles for main body of page
Set the styles for the sidebar elements. They weren't lining up right,
so I changed the main grid to be 4 rows, so that each section head
would be in its own row. Sizing of everything is still a little off,
but I just need to fix that and add some border shadows and I'll be
done.