Finish sizing grids
Made some changes to many of the grids. Biggest is to the "main"
section, as I turned it into a 4-column grid, with the projects
spanning three of the columns, and using grid-area to position the
sidebar elements. This allowed me to keep everything aligned more
easily.

I also changed the body-container, to limit the width of the nav
sidebar, as it was way too wide when the page was wide.

I also limited the width of the body, to keep the header aligned with
the main section. There has to be a better way to do this, and I will
explore this more later.